Adfectus 0.0.18
- Added functions to the Texture asset to flip horizontally and vertically.
  - This was previously done using the ModifyMatrix function but since the TextureMatrix concept was moved to the DesktopGL platform, something more generic must be used.
- Added a new math helper library - MathFloat
 - It includes some helpers for Float-Int interop, and C# Math functions with float arguments, but also some useful unique functions relating to angles.
- Moved the old helper-like files from Adfectus.Common and Adfectus.Primitives to a new namespace - Adfectus.Utility
- Added a new DebugMode property on the Engine class - as more debug functionality is moving to runtime rather than compile time.
- Added a QuadTree implementation
- Fixed the DefaultLogger not logging to stdOut when DebugMode was on.
- Added a 2D Ray/Line class to primitives.
- Added LOADS of new Rectangle functions.
 - Also introduced fixed to older Rectangle functions such as the "Intersects" function.
- Added a new "Bounds" property to the Transform class representing the rectangle bounds of the transform.
- Added a variety of new functions to the MathExtension class.
- Applied Code Style

+++ Adfectus/Game/SaveFile.cs
@@ -3,7 +3,7 @@
 using System;
 using System.IO;
 using System.Xml.Serialization;
-using Adfectus.Common;
+using Adfectus.Utility;
 
 #endregion
 
@@ -55,10 +55,7 @@ namespace Adfectus.Game
         public void Save()
         {
             // If an old save file exists, back it up.
-            if (File.Exists(_path))
-            {
-                File.Copy(_path, _path + ".backup", true);
-            }
+            if (File.Exists(_path)) File.Copy(_path, _path + ".backup", true);
 
             using (StreamWriter stream = new StreamWriter(_path))
             {
